up to date

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: tamigore <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/08 18:36:24 by tamigore          #+#    #+#             */
-/*   Updated: 2019/08/04 13:22:59 by tamigore         ###   ########.fr       */
+/*   Updated: 2019/11/11 19:32:22 by tamigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
